requirements.txt
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0
pyyaml>=6.0.0
geopandas>=0.14.0
folium>=0.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
